fixed bug in background subtraction
There is a silent bug in the Background Subtraction method. The matlab function graythresh, which uses Otsu's method for find a threshold, often just returns 0, resulting in the background intensity being calculated using all of the image. Microbetracker does not warn the user if this occurs. The mask then includes everything, defeating the whole purpose of making a mask. This results in too high of a background intensity being subtracted form the fluorescent image, thereby increasing the noise to signal ratio. This would be especially problematic for high cell densities.

I don't know under which conditions graythresh fails but I have found the problem in every one of the stacks I have checked so far. Usually, the effect is not noticeable but sometimes it results in erratic jumps in the whole cell and spot intensity profiles. I noticed these in some stacks several months ago but could not identity the problem.

A temporary fix is too expand the range of the image before finding the threshold. The algorithm then works and finds the correct (for that algorithm) threshold value. My phase contrast images only use a small proportion of the 16bit dynamic range, which seems to cause problems.
In subtractbgr.m line 8
replace
imgP = rawPhaseData(:,:,i);
with
imgP = imadjust(rawPhaseData(:,:,i));

One should probably also insert a warning if thres==0.

A better fix would be not to use this algorithm at all. Compare Otsu's method to Intermodes and Isofit in the Threshold plugin of imageJ. Both are usually much better at finding a good threshold value.

The above fix was was for the stand-alone function subtractbgr.m
For the Microtracker GUI do the following:
In microbeTracker.m line 3513 (in function subtractbgr)
replace
            if imsizes(1,3)==1
                imgP =rawPhaseData(:,:,1);
            else
                imgP = rawPhaseData(:,:,i);
            end
with
            if imsizes(1,3)==1
                imgP = imadjust(rawPhaseData(:,:,1));
            else
                imgP = imadjust(rawPhaseData(:,:,i));
            end

And in line 3526
replace
thres = graythreshreg(imgP,p.threshminlevel);
with
thres = graythreshreg(imgP);

This is because the previous change above (imadjust) expands the dynamic range so that the darkest pixels (the cells) are 0 and the brightest, 1. Therefore, we must must prevent the program using a minimum threshold ('thresminlevel') otherwise we would be excluding the cells in the calculation of the threshold.

With these two changes, background subtraction in the GUI does the same thing as the (modified) stand-alone function subtractbgr.m

It would still be better, however, to use a different method to calculate the threshold since the method above still chooses the threshold inappropriately sometimes.